+++ src/CommandLine/DocumentationHost.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) Josef Pihrt and Contributors. Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace Roslynator.CommandLine
+{
+    internal enum DocumentationHost
+    {
+        GitHub,
+        Docusaurus,
+    }
+}