+++ src/Documentation/CommonDocumentationParts.cs
@@ -1,9 +1,13 @@
 ﻿// Copyright (c) Josef Pihrt and Contributors. Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
+
 namespace Roslynator.Documentation
 {
-    public static class WellKnownUrlProviders
+    [Flags]
+    public enum CommonDocumentationParts
     {
-        public static DocumentationUrlProvider GitHub { get; } = GitHubDocumentationUrlProvider.Instance;
+        None = 0,
+        Content = 1,
     }
 }